2.17.72.1 - Fixes after Azure Cosmos implementation (#13)
* Fixed Enum types not read from Front.

* Fix drawing date not being saved

* Fix drawings not sorted by popularity when admin logged

* 2.17.72.1

+++ Tests/MRA.Services.Tests/Models/Drawings/DrawingServiceTests.cs
@@ -453,10 +453,10 @@ public class DrawingServiceTests : DocumentModelServiceTestsBase
             Name = name,
             ModelName = modelName,
             Title = title,
-            Software = DrawingSoftwares.ClipStudioPaint,
-            Paper = DrawingPaperSizes.A4,
-            Type = DrawingTypes.Digital,
-            ProductType = DrawingProductTypes.Videogame,
+            Software = (int) DrawingSoftwares.ClipStudioPaint,
+            Paper = (int) DrawingPaperSizes.A4,
+            Type = (int) DrawingTypes.Digital,
+            ProductType = (int) DrawingProductTypes.Videogame,
             ProductName = productName,
         };
